feat: implement update-refundsingle-token-transfer-logic-for-smart-contract with tests and docs
- Expand refund_single_token.rs with module NatSpec, security assumptions, and
  inline docs; delegate module refund_single transfer to refund_single_transfer.
- Merge refund_single_token_tests + refund_single_token_test into
  refund_single_token.test.rs (internal_token_helpers submodule); wire via
  #[path] in lib.rs; remove duplicate/orphan files and fix corrupted .test.rs.
- Update refund_single_token.md (correct CEI wording, module table, coverage);
  root refund_single_token.md points to src doc.
